Tests/LibWeb/Text/expected/SVG/use-element-inserted-together-with-referenced-element.txt

Referenced element before use: has instance root: true
Use before referenced element: has instance root: true
